question mark seems to work
600 lines of (seemingly) perfectly working code despite a fucking
depressing shadowing variable bug and relative complexity is not a bad
day. I have certainly gotten pretty good at to old programming.

Plus I am really enjoying this kernel coding style. It is actually very
pleasant: fast to write and fast to read. Helps you to lay it out in
your head and keep track of the complexity when you can see all your
variables at the top. It forces you to come up with a clean solution
with appropriate code, as opposed to letting you sneak in bad hacks and
cheats, which just worsens understanding and subtely increases the
complexity of what you have written.